% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_endpoint.R
\name{get_endpoint}
\alias{get_endpoint}
\title{Queries an Endpoint and returns a DataFrame}
\usage{
get_endpoint(endpoint, params = list())
}
\arguments{
\item{endpoint}{The specific endpoint to query. For example \code{"Loci"}.}

\item{params}{A named list of query parameters to include in the request.}
}
\value{
A character vector of length 1.
}
\description{
The \code{get_endpoint} function queries a STRBase API endpoint and retrieves data.
}
\examples{
\dontrun{
  # Basic query
  result <- get_endpoint(
    endpoint = "Loci")
  )

  # Print the result
  print(result)
}
}
\seealso{
\link{get_endpoint_JSON} for a version that returns the raw JSON response.
}
