% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locus_category_table.R
\name{get_locus_category_table}
\alias{get_locus_category_table}
\title{Retrieves the Locus Category table}
\usage{
get_locus_category_table(return_raw_data = FALSE)
}
\arguments{
\item{return_raw_data}{When \code{FALSE} (default), a selection of variables will be returned consistent with the web interface. When \code{TRUE}, all raw data available through the API will be returned.}
}
\value{
A DataFrame.
}
\description{
The \code{get_locus_category_table} function queries the STRbase API and retrieves the Locus Category table.
}
\examples{
\dontrun{
  # Basic query
  result <- get_locus_category_table()

  # Print the result
  print(result)
}
}
