% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locus_information.R
\name{get_locus_information}
\alias{get_locus_information}
\title{Retrieves Basic Information for a locus}
\usage{
get_locus_information(locus)
}
\arguments{
\item{locus}{The specific locus. For example \code{"FGA"}.}
}
\value{
A List.
}
\description{
The \code{get_locus_information} function queries the STRbase API and retrieves basic information for a locus. This includes the \code{lociId} which can be used when querying other endpoints.
}
\examples{
\dontrun{
  # Basic query
  result <- get_locus_information("FGA")

  # Print the result
  print(result)
}
}
