% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_locus_variant_alleles_table.R
\name{get_locus_variant_alleles_table}
\alias{get_locus_variant_alleles_table}
\title{Retrieves the Variant Alleles table for a locus}
\usage{
get_locus_variant_alleles_table(locus)
}
\arguments{
\item{locus}{The specific locus. For example \code{"FGA"}.}
}
\value{
A DataFrame.
}
\description{
The \code{get_locus_variant_alleles_table} function queries the STRBase API and retrieves the Variant Alleles table for a locus.
}
\examples{
\dontrun{
  # Basic query
  result <- get_locus_variant_alleles_table("FGA")

  # Print the result
  print(result)
}
}
