% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRbaseclient.R
\name{STRbaseclient}
\alias{STRbaseclient}
\alias{STRbaseclient-package}
\title{STRbaseclient}
\description{
Access NIST's STRbase API.
}
\details{
NIST's STRBase hosts STR-related information for the human identity community. Data can be browsed through "https://strbase.nist.gov/". This client accesses the API directly. This makes it easy to use the data from STRbase in your own analyses.
}
\seealso{
\itemize{
\item \url{https://strbase.nist.gov} for the STRbase website.
}
}
