% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STRBaseclient.R
\docType{package}
\name{STRBaseclient-package}
\alias{STRBaseclient}
\alias{STRBaseclient-package}
\title{STRBaseclient: Access NIST's STRBase API}
\description{
NIST's STRBase hosts STR-related information for the human identity community. Data can be browsed through "https://strbase.nist.gov/". This client accesses the API directly.

Access NIST's STRBase API.
}
\details{
NIST's STRBase hosts STR-related information for the human identity community. Data can be browsed through "https://strbase.nist.gov/". This client accesses the API directly. This makes it easy to use the data from STRBase in your own analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mkruijver/STRBaseclient}
  \item Report bugs at \url{https://github.com/mkruijver/STRBaseclient/issues}
}


\itemize{
\item \url{https://strbase.nist.gov} for the STRBase website.
}
}
\author{
\strong{Maintainer}: Maarten Kruijver \email{maarten.kruijver@esr.cri.nz} (\href{https://orcid.org/0000-0002-6890-7632}{ORCID})

}
\keyword{internal}
